+++ packages/core/src/types/RotatePageEvent.ts
@@ -1,0 +1,15 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2022 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+import type { PdfJs } from './PdfJs';
+
+export interface RotatePageEvent {
+    doc: PdfJs.PdfDocument;
+    pageIndex: number;
+    rotation: number;
+}